fix: rename and docs
