+++ packages/core/client/src/data-source/collection-fields-initializer-items/index.ts
@@ -1,3 +1,12 @@
-export * from './CollectionFields';
-export * from './FormCollectionFields';
-export * from './TableCollectionFields';
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+export * from './CollectionFieldsToInitializerItems';
+export * from './CollectionFieldsToFormInitializerItems';
+export * from './CollectionFieldsToTableInitializerItems';
