+++ packages/core/client/src/modules/blocks/data-blocks/form/formItemInitializers.tsx
@@ -7,7 +7,6 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import { FormCollectionFields } from '../../../../data-source';
 import { CompatibleSchemaInitializer } from '../../../../application/schema-initializer/CompatibleSchemaInitializer';
 import { AssociatedFields, ParentCollectionFields } from '../../../../schema-initializer/buttons/FormItemInitializers';
 import { gridRowColWrap, useFormItemInitializerFields } from '../../../../schema-initializer/utils';
@@ -17,10 +16,6 @@ const commonOptions = {
   icon: 'SettingOutlined',
   title: '{{t("Configure fields")}}',
   items: [
-    {
-      name: 'collectionFields',
-      Component: FormCollectionFields,
-    },
     {
       type: 'itemGroup',
       name: 'displayFields',
