+++ packages/core/client/src/modules/blocks/data-blocks/table/TableColumnInitializers.tsx
@@ -18,7 +18,6 @@ import {
   useInheritsTableColumnInitializerFields,
   useTableColumnInitializerFields,
 } from '../../../../schema-initializer/utils';
-import { TableCollectionFields } from '../../../../data-source/collection-fields-initializer-items';
 
 // 表格列配置
 const ParentCollectionFields = () => {
@@ -80,10 +79,6 @@ const commonOptions = {
     };
   },
   items: [
-    {
-      name: 'collectionFields',
-      Component: TableCollectionFields,
-    },
     {
       name: 'displayFields',
       type: 'itemGroup',
